💄 Refactor js files to common module
