+++ src/main/resources/static/js/common/modal.js
@@ -1,6 +1,7 @@
-// modal.js
+import { saveMovie, deleteWish } from "./api.js";
+
 export function showMovieModal(movie) {
-  const posters = movie.movieDetail.posters.filter((poster) => poster); // 빈 문자열 필터링
+  const posters = movie.detail.posters.filter((poster) => poster); // 빈 문자열 필터링
   const hasMultiplePosters = posters.length > 1;
   const posterCarousel = hasMultiplePosters
     ? `
@@ -40,7 +41,7 @@ export function showMovieModal(movie) {
           </div>
           <div class="modal-body">
             ${posterCarousel}
-            <p id="modalPlot">${movie.movieDetail.plot}</p>
+            <p id="modalPlot">${movie.detail.plot}</p>
           </div>
           <div class="modal-footer">
             <button type="button" class="btn btn-outline-primary" id="saveButton">저장</button>
@@ -70,7 +71,7 @@ export function showMovieModal(movie) {
 }
 
 export function showWishModal(wish) {
-  const posters = wish.movieDetail.posters.filter((poster) => poster); // 빈 문자열 필터링
+  const posters = wish.detail.posters.filter((poster) => poster); // 빈 문자열 필터링
   const hasMultiplePosters = posters.length > 1;
   const posterCarousel = hasMultiplePosters
     ? `
@@ -110,7 +111,7 @@ export function showWishModal(wish) {
           </div>
           <div class="modal-body">
             ${posterCarousel}
-            <p id="modalPlot">${wish.movieDetail.plot}</p>
+            <p id="modalPlot">${wish.detail.plot}</p>
           </div>
           <div class="modal-footer">
             <button type="button" class="btn btn-outline-primary" id="deleteButton" data-wish-id="${wish.id}">삭제</button>
